learn/requirements.txt
cachetools==5.2.0
db-dtypes==1.0.4
django-classy-tags==4.0.0
django-cms==3.11.0
django-encrypted-field==1.0.4
django-formtools==2.4
django-sekizai==4.0.0
django-treebeard==4.5.1
djangocms-admin-style==3.2.0
djangocms-simple-accordion==0.1.0
fonttools==4.25.0
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.7.1
google-cloud-bigquery==3.4.0
google-cloud-bigquery-storage==2.16.2
google-cloud-core==2.3.2
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
munkres==1.1.4
mysql-connector-python==8.0.31
oauthlib==3.2.2
pandas==1.4.4
pandas-gbq==0.18.1
Pillow==9.2.0
ply==3.11
proto-plus==1.22.1
protobuf==3.20.1
psycopg2-binary==2.9.5
pyarrow==9.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodomex==3.16.0
pydata-google-auth==1.4.0
PyMySQL==1.0.2
PyQt5-sip==12.11.0
requests-oauthlib==1.3.1
rsa==4.9
seaborn==0.12.1
uritemplate==4.1.1
whitenoise==6.2.0
